backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0

# Database
sqlalchemy==2.0.35
asyncpg==0.30.0
alembic==1.13.0

# Embeddings & Vector Search (FREE, local)
sentence-transformers==3.4.1
faiss-cpu==1.13.2
numpy==2.4.4

# LLM (Groq free tier primary, OpenAI fallback)
groq==0.11.0
openai==1.50.0

# Resume parsing
pdfplumber==0.11.0
python-docx==1.1.0

# Job scraping (public pages only)
httpx==0.27.0
beautifulsoup4==4.12.3
aiohttp==3.10.0


# Notifications
python-telegram-bot==21.5

# Utilities
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
apscheduler==3.10.4
tenacity==9.0.0
